+++ CHANGELOG.md
@@ -85,5 +85,7 @@
 - MVP-526 - Add Learn introduction page with slider
 - MVP-433 - Fix last step of signin 404
 - no-ticket - Fix images on /learn/categories being hidden
+- MVP-438 - Only view learn interstitials once 
+- MVP-464 - Store selected product as expertise
 - MVP-420 - test tooltips for multiple countries on target markets data page
 - MVP-438 - Only view learn interstitials once